+++ R/manual-edit.R
@@ -4,17 +4,22 @@
 #'
 #' `r lifecycle::badge("stable")`
 #'
-#' Needed for "manual edit" RStudio 'Addin' functionality.
+#' Needed for "manual edit" RStudio 'Addin' functionality.  Performs a git
+#' reset, git adds and commits a temporary version of control file to index (for
+#' subsequent diff), opens editor for edits.  The intent is that after the diff
+#' is created the commit will be removed.
 #'
 #' @param m An nm object
-#' @param combine_patch Logical. Should patch be added first before edit.
+#' @param combine_patch Character. Optional patch to be applied first.
+#' @param replace_ctl If not `NA_character_`, before displaying, the file will
+#'   be replaced with this
 #'
 #' @return A `list` with the temporary control file name and patch details (name
 #'   and path).
 #'
 #' @keywords internal
 #' @export
-start_manual_edit <- function(m, combine_patch = NA_character_) {
+start_manual_edit <- function(m, combine_patch = NA_character_, replace_ctl = NA_character_) {
   # if(.Platform$OS.type != "unix")
   #   stop("patching functionality only implemented for linux/unix systems\n consider manual_edit() instead",
   #        call. = FALSE)
@@ -25,29 +30,39 @@ start_manual_edit <- function(m, combine_patch = NA_character_) {
 
   time_stamp <- format(Sys.time(), "%Y-%m-%d-%H-%M-%S")
 
-  patch_name <- paste0("patch-", Sys.info()["user"], "-", time_stamp)
-  patch_path <- file.path(nm_default_dir("models"), "patches", patch_name)
+  patch_id <- paste0(Sys.info()["user"], "-", time_stamp)
+  patch_name <- paste0("patch-", patch_id)
+  patch_path <- file.path(nm_dir("models"), "patches", patch_name)
   dir.create(dirname(patch_path), showWarnings = FALSE, recursive = TRUE)
 
-  temp_ctl_path <- file.path(run_in(m), paste0("manual_", ctl_name(m)))
+  #temp_ctl_path <- file.path(run_in(m), paste0("manual_", ctl_name(m)))
+  temp_ctl_path <- file.path(nm_dir("models"), paste0("base-", patch_id))
   mnew <- m %>%
     ctl_path(temp_ctl_path) %>%
     write_ctl(force = TRUE)
 
   ## soft unstage all first, then add only the file
   if (!git_cmd_available) stop("need git available from system() for this to work")
-
   if (!user_values_exist()) stop("git user.name and/or user.email not set")
 
   system("git reset", intern = TRUE) ## for some reason git2r::reset() doesn't reset
   git2r::add(path = ctl_path(mnew))
-  git2r::commit(message = paste("before manual change: ", ctl_path(mnew)))
+  git2r::commit(message = paste("pre-manual edit: ", patch_id))
 
-  if (!is.na(combine_patch)) m <- m %>% apply_manual_edit(combine_patch)
-
-  usethis::edit_file(ctl_path(mnew)) ## edit new one
+  if (!is.na(combine_patch)) {
+    mnew <- mnew %>% 
+      apply_manual_edit(combine_patch) %>%
+      write_ctl(force = TRUE)
+  }
+  
+  if (!identical(replace_ctl, NA_character_)) {
+    writeLines(replace_ctl, ctl_path(mnew))
+  }
+  
+  usethis::ui_silence(usethis::edit_file(ctl_path(mnew))) ## edit new one
 
   res <- list()
+  res$patch_id <- patch_id
   res$new_ctl_path <- temp_ctl_path
   res$patch_name <- patch_name
   res$patch_path <- patch_path
@@ -55,9 +70,44 @@ start_manual_edit <- function(m, combine_patch = NA_character_) {
   res
 }
 
+git_user_name_exists <- function() {
+  config_names <- c(names(git2r::config()$global), names(git2r::config()$local))
+  "user.name" %in% config_names
+}
+
+git_user_email_exists <- function() {
+  config_names <- c(names(git2r::config()$global), names(git2r::config()$local))
+  "user.email" %in% config_names
+}
+
 user_values_exist <- function() {
-  all(c("user.email", "user.name") %in% names(git2r::config()$global)) |
-    all(c("user.email", "user.name") %in% names(git2r::config()$local))
+  config_names <- c(names(git2r::config()$global), names(git2r::config()$local))
+  all(c("user.email", "user.name") %in% config_names)
+}
+
+check_git_uservalues <- function() {
+  
+  cmd <- NA_character_
+  
+  if (!git_user_name_exists()) {
+    txt <- "user.name"
+    cmd <- "usethis::use_git_config(user.name = 'Jane')"
+  }
+  if (!git_user_email_exists()) {
+    txt <- "user.email"
+    cmd <- "usethis::use_git_config(user.email = 'jane@example.com')"
+  }
+  if (!git_user_name_exists() & !git_user_email_exists()) {
+    txt <- "user.name & user.email"
+    cmd <- "usethis::use_git_config(user.name = 'Jane', user.email = 'jane@example.com')"
+  }
+  
+  if(is.na(cmd)) return(invisible())
+  
+  usethis::ui_stop(paste0("git ", txt, "not set up. Set up like so:\n   ", 
+                          "{usethis::ui_code(\"", cmd, "\")}",
+                          "\nto set up"))
+  
 }
 
 check_git_username <- function() {
@@ -93,10 +143,18 @@ a user.name and user.email, set this up with:
 }
 
 diff_manual_edit <- function(m, res) {
+  
   git2r::add(path = res$new_ctl_path)
-  git2r::diff(git2r::repository(), index = TRUE, as_char = TRUE, filename = res$patch_path)
+  ## git2r::reset() doesn't work for some reason
+  on.exit(system("git reset HEAD", intern = TRUE))
+  
+  diff_text <- git2r::diff(git2r::repository(), index = TRUE, as_char = TRUE)
+  diff_text <- strsplit(diff_text, "\n")[[1]]
+  writeLines(diff_text, res$patch_path)
+  
+  ## following doesn't work for some reason
+  #git2r::diff(git2r::repository(), index = TRUE, as_char = TRUE, filename = res$patch_path)
   ## remove last commit and file
-  system("git reset HEAD^1", intern = TRUE) ## for some reason git2r::reset() doesn't reset
   unlink(res$new_ctl_path)
 }
 
@@ -105,15 +163,20 @@ diff_manual_edit <- function(m, res) {
 #'
 #' Use the "view patch" RStudio 'Addin' for viewing patches instead.
 #'
-#' @param patch_name Character. Name of patch
+#' @param patch_id Character. Name of patch
 #'
 #' @return No return value, called for side effects.
 #'
 #' @keywords internal
 #' @export
-view_patch <- function(patch_name) {
-  patch_path <- file.path(nm_default_dir("models"), "patches", patch_name)
-  file.show(patch_path)
+view_patch <- function(patch_id) {
+
+  patch_name <- paste0("patch-", patch_id)
+  patch_path <- file.path(nm_dir("models"), "patches", patch_name)
+  if (!file.exists(patch_path)) stop("patch file doesn't exist")
+  patch_contents <- readLines(patch_path)
+  cat(patch_contents, sep = "\n")
+  
 }
 
 new_patch_app <- function() {
@@ -124,16 +187,26 @@ new_patch_app <- function() {
 
   m <- get_single_object_for_app()
 
+  ## before doing start_manual_edit - log the current commit for later resetting
+  if (!git_cmd_available) stop("need git available from system() for this to work")
+  if (!user_values_exist()) stop("git user.name and/or user.email not set")
+  git_log <- git2r::commits()
+  orig_commit <- git_log[[1]]
+  
   res <- start_manual_edit(m)
+  on.exit(unlink(res$new_ctl_path))
+  
+  ans <- usethis::ui_yeah("---INSTRUCTIONS--- 
 
-  message(
-    "---Manual edit---
-    Instructions:
-    1) edit control file
-    2) save & close
-    Press ENTER when done..."
-  )
-  readline()
+ 1) {usethis::ui_field('EDIT')} control file
+ 2) {usethis::ui_field('SAVE')} the file
+ 
+Finished & happy to proceed?", yes = "Yes", no = "Abort", shuffle = FALSE)
+  
+  if(!ans) {
+    system(paste("git reset", orig_commit$sha), intern = TRUE)
+    return(invisible())
+  }
 
   ## now diff ctl_path(m) and old_file_path
 
@@ -141,14 +214,14 @@ new_patch_app <- function() {
 
   if (final_pipe_present) {
     if (final_newline_present) {
-      code_to_add <- paste0("  apply_manual_edit(\"", res$patch_name, "\")")
+      code_to_add <- paste0("  apply_manual_edit(\"", res$patch_id, "\") %>%")
     } else {
       code_to_add <- paste0("
-  apply_manual_edit(\"", res$patch_name, "\")")
+  apply_manual_edit(\"", res$patch_id, "\") %>%")
     }
   } else {
     code_to_add <- paste0(" %>%
-  apply_manual_edit(\"", res$patch_name, "\")")
+  apply_manual_edit(\"", res$patch_id, "\")")
   }
 
 
@@ -168,20 +241,30 @@ modify_patch_app <- function() {
 
   before_edit <- gsub("(.*)%>%.*apply_manual_edit.*", "\\1", selected_text)
 
-  patch_name <- gsub(".*%>%.*apply_manual_edit\\(\"(.*)\"\\).*", "\\1", selected_text)
+  patch_id <- gsub(".*%>%.*apply_manual_edit\\(\"(.*)\"\\).*", "\\1", selected_text)
 
   m <- eval(parse(text = before_edit))
+  
+  ## before doing start_manual_edit - log the current commit for later resetting
+  if (!git_cmd_available) stop("need git available from system() for this to work")
+  if (!user_values_exist()) stop("git user.name and/or user.email not set")
+  git_log <- git2r::commits()
+  orig_commit <- git_log[[1]]
+  
+  res <- start_manual_edit(m, combine_patch = patch_id)
+  on.exit(unlink(res$new_ctl_path))
+  
+  ans <- usethis::ui_yeah("---INSTRUCTIONS--- 
 
-  res <- start_manual_edit(m, combine_patch = patch_name)
-
-  message(
-    "---Manual edit---
-    Instructions:
-    1) edit control file
-    2) save & close
-    Press ENTER when done..."
-  )
-  readline()
+ 1) {usethis::ui_field('EDIT')} control file
+ 2) {usethis::ui_field('SAVE')} the file
+ 
+Finished & happy to proceed?", yes = "Yes", no = "Abort", shuffle = FALSE)
+  
+  if(!ans) {
+    system(paste("git reset", orig_commit$sha), intern = TRUE)
+    return(invisible())
+  }
 
   ## now diff ctl_path(m) and old_file_path
 
@@ -189,14 +272,14 @@ modify_patch_app <- function() {
 
   if (final_pipe_present) {
     if (final_newline_present) {
-      code_to_add <- paste0(trimws(before_edit), "  apply_manual_edit(\"", res$patch_name, "\") %>%")
+      code_to_add <- paste0(trimws(before_edit), "  apply_manual_edit(\"", res$patch_id, "\") %>%")
     } else {
       code_to_add <- paste0(trimws(before_edit), " %>%
-  apply_manual_edit(\"", res$patch_name, "\") %>%")
+  apply_manual_edit(\"", res$patch_id, "\") %>%")
     }
   } else {
     code_to_add <- paste0(trimws(before_edit), " %>%
-  apply_manual_edit(\"", res$patch_name, "\")")
+  apply_manual_edit(\"", res$patch_id, "\")")
   }
 
   rstudioapi::insertText(
@@ -207,8 +290,73 @@ modify_patch_app <- function() {
   message("apply_manual_edit() statement modified in script")
 }
 
+resolve_manual_edit <- function() {
+  ctx <- rstudioapi::getSourceEditorContext()
+  selected_text <- ctx$selection[[1]]$text
+  final_pipe_present <- grepl("\\s*%>%\\s*$", selected_text)
+  final_newline_present <- grepl("\\n\\s*$", selected_text)
+
+  before_edit <- gsub("(.*)%>%.*apply_manual_edit.*", "\\1", selected_text)
+  
+  patch_id <- gsub(".*%>%.*apply_manual_edit\\(\"(.*)\"\\).*", "\\1", selected_text)
+  
+  m <- eval(parse(text = before_edit))
+
+  ctl_with_merge <- m %>%
+    apply_manual_edit(patch_id, return_merge_conf_ctl = TRUE)
+  
+  ctl_with_merge <- ctl_with_merge[[1]]
+  
+  ## before doing start_manual_edit - log the current commit for later resetting
+  if (!git_cmd_available) stop("need git available from system() for this to work")
+  if (!user_values_exist()) stop("git user.name and/or user.email not set")
+  git_log <- git2r::commits()
+  orig_commit <- git_log[[1]]
+  
+  res <- start_manual_edit(m, replace_ctl = ctl_with_merge)
+  on.exit(unlink(res$new_ctl_path))
+  
+  ans <- usethis::ui_yeah("---INSTRUCTIONS--- 
+
+ 1) {usethis::ui_field('REPAIR')} control file
+    a) Merge conflicts are indicated by the markers <<<<<<<, =======, and >>>>>>>
+    b) {usethis::ui_field('EDIT')} these sections
+ 2) {usethis::ui_field('SAVE')} the file
+ 
+Finished & happy to proceed?", yes = "Yes", no = "Abort", shuffle = FALSE)
+  
+  if(!ans) {
+    system(paste("git reset", orig_commit$sha), intern = TRUE)
+    return(invisible())
+  }
+  
+  ## now diff ctl_path(m) and old_file_path
+  
+  diff_manual_edit(m, res)
+  
+  if (final_pipe_present) {
+    if (final_newline_present) {
+      code_to_add <- paste0(trimws(before_edit), "  apply_manual_edit(\"", res$patch_id, "\") %>%")
+    } else {
+      code_to_add <- paste0(trimws(before_edit), " %>%
+  apply_manual_edit(\"", res$patch_id, "\") %>%")
+    }
+  } else {
+    code_to_add <- paste0(trimws(before_edit), " %>%
+  apply_manual_edit(\"", res$patch_id, "\")")
+  }
+  
+  rstudioapi::insertText(
+    text = code_to_add,
+    id = ctx$id
+  )
+  
+  message("apply_manual_edit() statement modified in script")
+  
+}
+
 make_patch_app <- function() {
-  check_git_username()
+  check_git_uservalues()
 
   ctx <- rstudioapi::getSourceEditorContext()
   selected_text <- ctx$selection[[1]]$text
@@ -237,6 +385,38 @@ make_patch_app <- function() {
   }
 }
 
+resolve_manual_edit_app <- function() {
+  check_git_uservalues()
+  
+  ctx <- rstudioapi::getSourceEditorContext()
+  selected_text <- ctx$selection[[1]]$text
+  selected_text <- gsub("\\s*%>%\\s*$", "", selected_text)
+  ## see if last function is apply_manual_edit
+  
+  ## get last function used in piping
+  full_expr <- parse(text = selected_text)[[1]]
+  
+  if (as.character(full_expr[[1]]) == "<-") {
+    full_expr <- full_expr[[3]] ## rhs of assignment
+  }
+  
+  if (as.character(full_expr[[1]]) == "%>%") {
+    last_fun_name <- as.character(full_expr[[3]][[1]])
+  }
+  
+  if (as.character(full_expr[[1]]) != "%>%") { ## normal function
+    last_fun_name <- as.character(full_expr[[1]])
+  }
+  
+  if (last_fun_name == "apply_manual_edit") {
+    resolve_manual_edit()
+  } else {
+    stop("the last function applied to selection needs to be apply_manual_edit()")
+  }
+}
+
+
+
 
 view_patch_app <- function() {
   ctx <- rstudioapi::getActiveDocumentContext()
@@ -246,7 +426,7 @@ view_patch_app <- function() {
     selected_text <- ctx$contents[line]
   }
 
-  selected_text <- gsub(".*(patch-[^\"]+)\".*", "\\1", selected_text)
+  selected_text <- gsub(".*\"([^\"]+)\".*", "\\1", selected_text)
 
   # selected_text <- gsub("\"","", selected_text)